GEODE-6425: Fixes region stat counter rollover
After fixing GEODE-6424 operations can execute fast enough that some
int counters roll over too frequently to provide meaningful information.
Converting to long counters delays this rollover, which allows for more
meaningful analysis.

+++ geode-core/src/integrationTest/java/org/apache/geode/cache/ProxyJUnitTest.java
@@ -642,13 +642,13 @@ public class ProxyJUnitTest {
     ExpectedEntryEvent expee = new ExpectedEntryEvent();
     expee.r = r;
     expee.key = "key";
-    int creates = getStats().getCreates();
+    long creates = getStats().getCreates();
     // int puts = getStats().getPuts();
     // int updates = getStats().getUpdates();
-    int destroys = getStats().getDestroys();
-    int invalidates = getStats().getInvalidates();
-    int gets = getStats().getGets();
-    int misses = getStats().getMisses();
+    long destroys = getStats().getDestroys();
+    long invalidates = getStats().getInvalidates();
+    long gets = getStats().getGets();
+    long misses = getStats().getMisses();
 
     r.put("key", "value", cbArg);
     expee.op = Operation.CREATE;
@@ -836,13 +836,13 @@ public class ProxyJUnitTest {
     checkNoCW();
     checkCL(expre);
 
-    int creates = getStats().getCreates();
+    long creates = getStats().getCreates();
     // int puts = getStats().getPuts();
     // int updates = getStats().getUpdates();
-    int destroys = getStats().getDestroys();
+    long destroys = getStats().getDestroys();
     // int invalidates = getStats().getInvalidates();
-    int gets = getStats().getGets();
-    int misses = getStats().getMisses();
+    long gets = getStats().getGets();
+    long misses = getStats().getMisses();
     ExpectedEntryEvent expee = new ExpectedEntryEvent();
     expee.r = r;
     expee.key = "key";
@@ -928,11 +928,11 @@ public class ProxyJUnitTest {
     checkNoTL();
     checkCL(expre);
 
-    int creates = getStats().getCreates();
+    long creates = getStats().getCreates();
     // int puts = getStats().getPuts();
     // int updates = getStats().getUpdates();
-    int destroys = getStats().getDestroys();
-    int invalidates = getStats().getInvalidates();
+    long destroys = getStats().getDestroys();
+    long invalidates = getStats().getInvalidates();
     // int gets = getStats().getGets();
     // int misses = getStats().getMisses();
     ExpectedEntryEvent expee = new ExpectedEntryEvent();
